Move MD files to public folder
