+++ public/articles/vilperi.md
@@ -3,11 +3,11 @@
 GitHub  
 <a href="https://github.com/petterigit/petterigit.github.io" target="_blank">Website Git Link</a>
 
-<a href="https://www.typescriptlang.org/" target="_blank" rel="noreferrer"> ![Typescript](./articleElements/ts-logo-128.svg 'Typescript')</a>
-<a href="https://reactjs.org/" target="_blank" rel="noreferrer"> ![React](./articleElements/react-logo-128.png 'React')</a>
-<a href="https://sass-lang.com/" target="_blank" rel="noreferrer"> ![Sass](./articleElements/sass-logo-128.png 'Sass')</a>
-<a href="https://threejs.org/" target="_blank" rel="noreferrer"> ![ThreeJS](./articleElements/threejs-logo-128.png 'ThreeJS')</a>
-<a href="https://www.blender.org/" target="_blank" rel="noreferrer"> ![Blender](./articleElements/blender-logo-128.svg 'Blender')</a>
+<a href="https://www.typescriptlang.org/" target="_blank" rel="noreferrer"> ![Typescript](../articleElements/ts-logo-128.svg 'Typescript')</a>
+<a href="https://reactjs.org/" target="_blank" rel="noreferrer"> ![React](../articleElements/react-logo-128.png 'React')</a>
+<a href="https://sass-lang.com/" target="_blank" rel="noreferrer"> ![Sass](../articleElements/sass-logo-128.png 'Sass')</a>
+<a href="https://threejs.org/" target="_blank" rel="noreferrer"> ![ThreeJS](../articleElements/threejs-logo-128.png 'ThreeJS')</a>
+<a href="https://www.blender.org/" target="_blank" rel="noreferrer"> ![Blender](../articleElements/blender-logo-128.svg 'Blender')</a>
 
 ---
